+++ src/main/java/Duke.java
@@ -1,3 +1,4 @@
+import duke.*;
 import java.util.ArrayList;
 
 public class Duke {
