+++ src/main/java/duke/Task.java
@@ -1,3 +1,4 @@
+package duke;
 import java.io.Serializable;
 
 public class Task implements Serializable {
