+++ src/main/java/duke/TaskList.java
@@ -1,10 +1,11 @@
+package duke;
 import java.util.ArrayList;
 
 public class TaskList {
 
     protected ArrayList<Task> arr;
 
-    public TaskList(ArrayList<Task> arr) {
+    public TaskList(ArrayList<duke.Task> arr) {
         this.arr = arr;
     }
